Bugfixes nach Test - Globale SystemFlag-Variablen für Zykluszeitmessung werden nun mit Werten belegt - Header-Dateien PlcMemory.h und teranis.h vereinheitlicht, weil sowieso beide nur zusammen verwendet wurden
