+++ TeranisTask.cpp
@@ -1,14 +1,10 @@
 
 #include "PlcMemory.h"
-
-#include "teranis.h"
-;
-
 #include <cstdint>
+#include <iostream>
 #include "teranis_code.h"
-
-
 #include "TeranisTask.h"
+#include "avicstd/h/avi_c_std.h"
 
 TeranisTask::TeranisTask() : PlcTask() {
 }
